externalize options and values for easier overloading
